refactored to share groupsource/groupactions between the apps
