+++ app/src/main/admin/components/ModellingGroups/SingleGroup/Details/ViewModellingGroupDetailsPage.tsx
@@ -4,7 +4,7 @@ import { connectToStores } from "../../../../../shared/alt";
 import { ModellingGroupTitle, TitleProps } from "../ModellingGroupTitle";
 import { AdminPageWithHeader } from "../../../AdminPageWithHeader";
 import { groupStore } from "../../../../stores/GroupStore";
-import { modellingGroupActions } from "../../../../actions/ModellingGroupActions";
+import { modellingGroupActions } from "../../../../../shared/actions/ModellingGroupActions";
 import { doNothing } from "../../../../../shared/Helpers";
 import { userStore } from "../../../../stores/UserStore";
