+++ app/src/main/admin/stores/GroupStore.ts
@@ -2,9 +2,9 @@ import * as AltJS from "alt";
 import { ModellingGroup, ModellingGroupDetails } from "../../shared/models/Generated";
 import { alt } from "../../shared/alt";
 import { AbstractStore } from "../../shared/stores/AbstractStore";
-import { modellingGroupActions } from "../actions/ModellingGroupActions";
+import { modellingGroupActions } from "../../shared/actions/ModellingGroupActions";
 import { RemoteContent } from "../../shared/models/RemoteContent";
-import { ModellingGroupSource } from "../sources/ModellingGroupSource";
+import { ModellingGroupSource } from "../../shared/sources/ModellingGroupSource";
 import { ILookup } from "../../shared/models/Lookup";
 import StoreModel = AltJS.StoreModel;
