+++ app/src/main/contrib/components/ChooseAction/ChooseActionPage.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { ChooseActionContent } from "./ChooseActionContent";
-import { modellingGroupActions } from "../../actions/ModellingGroupActions";
 import { PageWithHeaderAndNav } from "../PageWithHeader/PageWithHeaderAndNav";
 import { responsibilityStore } from "../../stores/ResponsibilityStore";
 import { doNothing } from "../../../shared/Helpers";
+import {modellingGroupActions} from "../../../shared/actions/ModellingGroupActions";
 
 export interface LocationProps {
     groupId: string;
@@ -12,7 +12,7 @@ export interface LocationProps {
 export class ChooseActionPage extends PageWithHeaderAndNav<LocationProps> {
     componentDidMount() {
         responsibilityStore.fetchTouchstones().catch(doNothing);
-        modellingGroupActions.setCurrentModellingGroup(this.props.location.params.groupId);
+        modellingGroupActions.setCurrentGroup(this.props.location.params.groupId);
     }
 
     title() {
