--- app/src/main/contrib/sources/ModellingGroupSource.ts
@@ -1,17 +1,0 @@
-import { ModellingGroup } from "../../shared/models/Generated";
-import { Source } from "../../shared/sources/Source";
-import { MainState } from "../stores/MainStore";
-import { modellingGroupActions } from "../actions/ModellingGroupActions";
-import SourceModel = AltJS.SourceModel;
-
-export class ModellingGroupSource extends Source<MainState> {
-    fetchModellingGroups: () => SourceModel<ModellingGroup[]>;
-
-    constructor() {
-        super();
-        this.fetchModellingGroups = () => this.doFetch(_ => "/modelling-groups/", {
-            success: modellingGroupActions.update,
-            loading: modellingGroupActions.beginFetch
-        });
-    }
-}
