+++ app/src/main/contrib/stores/ContribAuthStore.ts
@@ -2,9 +2,9 @@ import alt from "../../shared/alt";
 import { LogInProperties } from "../../shared/actions/AuthActions";
 import { mainStore } from "./MainStore";
 import { ModellingGroup } from "../../shared/models/Generated";
-import { modellingGroupActions } from "../actions/ModellingGroupActions";
 import { AuthStateBase, AuthStore, AuthStoreBaseInterface } from "../../shared/stores/AuthStoreBase";
 import StoreModel = AltJS.StoreModel;
+import {modellingGroupActions} from "../../shared/actions/ModellingGroupActions";
 
 export interface ContribAuthState extends AuthStateBase {
     modellingGroupIds: string[];
@@ -27,7 +27,7 @@ class ContribAuthStore extends AuthStore<ContribAuthState, ContribAuthStoreInter
     constructor() {
         super();
         this.bindListeners({
-            handleModellingGroups: modellingGroupActions.update
+            handleModellingGroups: modellingGroupActions.updateGroups
         });
     }
