+++ app/src/test/contrib/components/ChooseAction/ChooseActionPageTests.tsx
@@ -24,7 +24,7 @@ describe("ChooseActionPage", () => {
             expect(fetchTouchstones.called).to.equal(true, "Expected responsibilityStore.fetchTouchstones to be called");
             afterWait(() => {
                 expectOneAction(spy, {
-                    action: "ModellingGroupActions.setCurrentModellingGroup",
+                    action: "ModellingGroupActions.setCurrentGroup",
                     payload: "gId"
                 });
             });
