+++ app/src/test/contrib/stores/ResponsibilityStoreTests.ts
@@ -16,9 +16,9 @@ import { responsibilityActions } from "../../../main/contrib/actions/Responsibil
 import { responsibilityStore } from "../../../main/contrib/stores/ResponsibilityStore";
 import { coverageSetActions } from "../../../main/contrib/actions/CoverageSetActions";
 import { coverageTokenActions } from "../../../main/contrib/actions/CoverageActions";
-import { modellingGroupActions } from "../../../main/contrib/actions/ModellingGroupActions";
 import { makeLookup } from "../../../main/contrib/stores/Loadable";
 import { ModellingGroup } from "../../../main/shared/models/Generated";
+import {modellingGroupActions} from "../../../main/shared/actions/ModellingGroupActions";
 const jwt = require("jsonwebtoken");
 
 describe("ResponsibilityStore", () => {
@@ -45,33 +45,33 @@ describe("ResponsibilityStore", () => {
         });
     });
 
-    it("modellingGroupActions.setCurrentModellingGroup sets current modelling group", () => {
+    it("modellingGroupActions.setCurrentGroup sets current modelling group", () => {
         const group = mockModellingGroup();
         alt.bootstrap(JSON.stringify({
             MainStore: { modellingGroups: makeLookup<ModellingGroup>([ group ]) }
         }));
-        modellingGroupActions.setCurrentModellingGroup(group.id);
+        modellingGroupActions.setCurrentGroup(group.id);
         expect(responsibilityStore.getState().currentModellingGroup).to.eql(group);
     });
 
     it("modellingGroupActions.update sets current modelling group if only a member of one", () => {
         // User has no membership of any group
         const group = mockModellingGroup();
-        modellingGroupActions.update([ group ]);
+        modellingGroupActions.updateGroups([ group ]);
         expect(responsibilityStore.getState().currentModellingGroup).to.be.null;
 
         // User has membership of multiple groups
         alt.bootstrap(JSON.stringify({
             ContribAuthStore: { modellingGroupIds: [ group.id, "another-id" ] }
         }));
-        modellingGroupActions.update([ group ]);
+        modellingGroupActions.updateGroups([ group ]);
         expect(responsibilityStore.getState().currentModellingGroup).to.be.null;
 
         // User has membership of just one group
         alt.bootstrap(JSON.stringify({
             ContribAuthStore: { modellingGroupIds: [ group.id ] }
         }));
-        modellingGroupActions.update([ group ]);
+        modellingGroupActions.updateGroups([ group ]);
         expect(responsibilityStore.getState().currentModellingGroup).to.eql(group);
     });
